+++ redwood-lazylayout-compose/src/commonMain/kotlin/app/cash/redwood/lazylayout/compose/LoadingStrategy.kt
@@ -17,16 +17,30 @@ package app.cash.redwood.lazylayout.compose
 
 public interface LoadingStrategy {
   /**
-   * Returns the most recent first index scrolled to. This is used to save the scroll position when
-   * the view is unloaded.
+   * Returns the index of the first item that is visible on screen. The item may be partially
+   * visible.
+   *
+   * This is used to save the scroll position when the view is unloaded.
+   *
+   * This may temporarily be larger than the total number of items in the model. This will occur if
+   * the number of items in the model shrinks.
    */
-  public val firstIndex: Int
+  public val firstVisibleIndex: Int
 
-  /** Perform a programmatic scroll to [firstIndex]. */
-  public fun scrollTo(firstIndex: Int)
+  /**
+   * Returns the index of the last item that is visible on screen. The item may be partially
+   * visible.
+   *
+   * This may temporarily be larger than the total number of items in the model. This will occur if
+   * the number of items in the model shrinks.
+   */
+  public val lastVisibleIndex: Int
+
+  /** Perform a programmatic scroll to [firstVisibleIndex]. */
+  public fun scrollTo(firstVisibleIndex: Int)
 
   /** React to a user-initiated scroll to the target range. */
-  public fun onUserScroll(firstIndex: Int, lastIndex: Int)
+  public fun onUserScroll(firstVisibleIndex: Int, lastVisibleIndex: Int)
 
   /**
    * Returns the range of items to render into the view tree. This should be a slice of